Add separate section to configure portal navigation (repository).
