+++ opengever/setup/cfgs/repository.cfg
@@ -10,6 +10,7 @@ pipeline =
     dex_schemaupdater
     blocklocalroles
     ac_local_roles
+    assignreporootnavigation
     repositoryfirst
     reindexobject
     savepoint
@@ -57,6 +58,9 @@ fields = python:['block_inheritance']
 
 [ac_local_roles]
 blueprint = collective.blueprint.jsonmigrator.ac_local_roles
+
+[assignreporootnavigation]
+blueprint = opengever.setup.assignreporootnavigation
 
 [repositoryfirst]
 blueprint = opengever.setup.repositoryfirst
